move getActiveTeamCountByOrgIds to helper func and add total meeting count
